+++ CHANGELOG.md
@@ -18,3 +18,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Converter Template and Conversion Template
 - Input and Output Format in Converter
 - Load converter from application context
+- Converter Selector